Add explicit float dtype to test tensors

+++ tests/metrics/functional/test_classification.py
@@ -274,7 +274,7 @@ def test_auc(x, y, expected):
                 [0, 0, 0.2, 15, 201, 26, 254],
                 [0, 0, 0.65, 35, 45, 210, 170],
                 [0, 0, 0.7, 210, 80, 295, 90]
-            ]),
+            ], dtype=torch.float),
             torch.tensor([
                 [0, 0, 100., 50., 205., 200.],
                 [0, 0, 85., 83., 225., 195.],
@@ -282,7 +282,7 @@ def test_auc(x, y, expected):
                 [0, 0, 10., 200., 15., 254.],
                 [0, 0, 30., 40., 234., 150.],
                 [0, 0, 210., 80., 295., 90.],
-            ]),
+            ], dtype=torch.float),
             0.5,
         ),
     ]
@@ -309,13 +309,13 @@ def test_object_detection_mean_average_precision_0(pred, target, iou_threshold, 
                 [1, 0, 0.9, 100, 100, 200, 200],
                 [0, 2, 0.9, 100, 100, 200, 200],
                 [2, 1, 0.9, 100, 100, 200, 200],
-            ]),
+            ], dtype=torch.float),
             torch.tensor([
                 [0, 1, 100., 100., 200., 200.],
                 [1, 0, 100., 100., 200., 200.],
                 [0, 2, 100., 100., 200., 200.],
                 [2, 1, 100., 100., 200., 200.],
-            ]),
+            ], dtype=torch.float),
             0.5,
         )
     ]
@@ -342,8 +342,8 @@ def test_object_detection_mean_average_precision_1(pred, target, iou_threshold, 
                 [1, 0, 0.9, 100, 100, 200, 200],
                 [0, 2, 0.9, 100, 100, 200, 200],
                 [2, 1, 0.9, 100, 100, 200, 200],
-            ]),
-            torch.tensor([[0, 3, 100, 100, 200, 200]]),
+            ], dtype=torch.float),
+            torch.tensor([[0, 3, 100, 100, 200, 200]], dtype=torch.float),
             0.5,
         )
     ]
@@ -359,7 +359,6 @@ def test_object_detection_mean_average_precision_1(pred, target, iou_threshold, 
 def test_object_detection_mean_average_precision_no_target(pred, target, iou_threshold, ap_calculation, expected_map):
     mean_average_precision = object_detection_mean_average_precision(pred, target, iou_threshold, ap_calculation)
     assert torch.allclose(mean_average_precision, expected_map)
-
 
 
 @pytest.mark.parametrize(['pred', 'target', 'expected'], [
